add clean and batch delete

+++ main.go
@@ -1,10 +1,13 @@
 package main
 
 import (
+	"bytes"
 	"context"
+	"encoding/json"
 	"fmt"
 	"log"
 	"math/rand"
+	"net/http"
 	"os"
 	"os/signal"
 	"path/filepath"
@@ -16,6 +19,7 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/profiles/latest/containerservice/mgmt/containerservice"
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2021-03-01/compute"
+	"github.com/Azure/go-autorest/autorest"
 	"github.com/Azure/go-autorest/autorest/azure/auth"
 	"github.com/Azure/go-autorest/autorest/to"
 	corev1 "k8s.io/api/core/v1"
@@ -30,6 +34,8 @@ import (
 var nodesToDeleteWatch = make(map[string]bool)
 var mutexWatch sync.Mutex // 使用互斥锁来保证 map 的并发安全
 var nodesToCheckCycle = make(map[string]bool)
+var url = fmt.Sprintf("https://management.azure.com/subscriptions/%s/resourceGroups/%s/providers/Microsoft.Compute/virtualMachineScaleSets/%s/delete?api-version=2024-03-01",
+	os.Getenv("AZURE_SUBSCRIPTION_ID"), os.Getenv("VMSS_RG"), os.Getenv("VMSS_NAME"))
 
 func main() {
 	log.Printf("Hello")
@@ -45,6 +51,18 @@ func main() {
 		log.Printf("Error converting COOLDOWN to int: %s\n", err)
 		return
 	}
+
+	fakecoolDownStr := os.Getenv("FAKENP_COOLDOWN")
+	fakecoolDown, err := strconv.Atoi(fakecoolDownStr)
+	if err != nil {
+		log.Printf("Error converting FAKENP_COOLDOWN to int: %s\n", err)
+		return
+	}
+	if fakecoolDown == 0 {
+		fakecoolDown = 180
+	}
+
+	log.Printf("FAKENP_COOLDOWN value: %d\n", fakecoolDown)
 
 	// Start watching nodes
 	go watchK8SEvents(clientset, vmssClient, stopCh)
@@ -71,6 +89,8 @@ func main() {
 					if err != nil {
 						log.Printf("Error checking and deleting nodes: %v\n", err)
 					}
+				case <-time.After(time.Duration(fakecoolDown) * time.Second):
+					updateFakeNP(aksClient, os.Getenv("AKS_RG"), os.Getenv("AKS_NAME"), os.Getenv("FAKE_NODEPOOL"))
 				}
 			}
 		}()
@@ -81,20 +101,69 @@ func main() {
 	log.Println("shutting down node watcher...")
 }
 
+func DeleteVMInstances(vmssClient compute.VirtualMachineScaleSetVMsClient, instanceIDs []string) error {
+	// 创建一个 authorizer
+	authorizer := vmssClient.Authorizer
+	// 创建请求体
+	body := struct {
+		InstanceIds []string `json:"instanceIds"`
+	}{
+		InstanceIds: instanceIDs,
+	}
+	bodyBytes, err := json.Marshal(body)
+	if err != nil {
+		return err
+	}
+
+	// 构建 HTTP 请求
+	req, err := http.NewRequest("POST", url, bytes.NewReader(bodyBytes))
+	if err != nil {
+		return err
+	}
+	req.Header.Set("Content-Type", "application/json")
+
+	// 使用 authorizer 为请求添加授权头部
+	preparer := autorest.CreatePreparer(
+		authorizer.WithAuthorization(),
+	)
+	req, err = preparer.Prepare(req)
+	if err != nil {
+		return err
+	}
+
+	// 创建一个 HTTP 客户端并发送请求
+	client := autorest.NewClientWithUserAgent("Azure-Go-SDK/2024-03-01")
+	resp, err := client.Do(req)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+
+	// 检查响应状态码
+	if resp.StatusCode != http.StatusOK && resp.StatusCode != http.StatusAccepted {
+		log.Printf("API call to delete VM instances returned status: %s", resp.Status)
+		return fmt.Errorf("API call to delete VM instances returned status: %s", resp.Status)
+	}
+
+	// 处理响应
+	// log.Printf("VMSS DeleteResponse status: %s\n", resp.Status)
+
+	return nil
+}
+
 func processNodesbyEvents(clientset *kubernetes.Clientset, vmssClient compute.VirtualMachineScaleSetVMsClient) {
 	labelKey := os.Getenv("LABEL_KEY")
 	labelValue := os.Getenv("LABEL_VALUE")
 	vmssRG := os.Getenv("VMSS_RG")
 	namespace := os.Getenv("NAMESPACE")
 	coolDownStr := os.Getenv("COOLDOWN")
-	coolDown, err := strconv.Atoi(coolDownStr)
-	if err != nil {
-		log.Printf("Pod Event: Error converting COOLDOWN to int: %s\n", err)
-		return
-	}
+	coolDown, _ := strconv.Atoi(coolDownStr)
+	isDeleteInBatchStr := os.Getenv("DELETE_IN_BATCH")
+	isDeleteInBatch, _ := strconv.ParseBool(isDeleteInBatchStr)
+
 	// vmssName := os.Getenv("VMSS_NAME")
 	for {
-
+		vmssInstances := []string{}
 		mutexWatch.Lock()
 		// nodesToDeleteWatch 赋值给新的 map，以便在迭代时删除元素
 		nodesToDelete := make(map[string]bool)
@@ -134,10 +203,11 @@ func processNodesbyEvents(clientset *kubernetes.Clientset, vmssClient compute.Vi
 				// mutexWatch.Lock()
 
 				// Drain Node
-				err = DrainNode(clientset, nodeName)
-				log.Printf("Process Pod Event: Drain node: %s Completed \n", nodeName)
+				log.Printf("Process Pod Event: Cordon node: %s\n", nodeName)
+				err = CordonNode(clientset, nodeName)
+				log.Printf("Process Pod Event: Cordon node: %s Completed \n", nodeName)
 				if err != nil {
-					log.Printf("Process Pod Event: Error draining node %s: %v\n", nodeName, err)
+					log.Printf("Process Pod Event: Error Cordon node %s: %v\n", nodeName, err)
 					// mutexWatch.Unlock()
 					continue
 				}
@@ -150,6 +220,7 @@ func processNodesbyEvents(clientset *kubernetes.Clientset, vmssClient compute.Vi
 					// mutexWatch.Unlock()
 					continue
 				}
+				vmssInstances = append(vmssInstances, instanceID)
 
 				// 从 Kubernetes 集群中删除节点
 				log.Printf("Pod Event: Deleting K8S node : %s\n", nodeName)
@@ -162,12 +233,14 @@ func processNodesbyEvents(clientset *kubernetes.Clientset, vmssClient compute.Vi
 				log.Printf("Pod Event: Node %s deleted from Kubernetes\n", nodeName)
 
 				// 删除VMSS实例
-				log.Printf("Pod Event: Deleting VMSS instance with ID: %s\n", instanceID)
-				_, err = vmssClient.Delete(context.TODO(), vmssRG, vmssName, instanceID, nil)
-				if err != nil {
-					log.Printf("Pod Event: Error deleting VMSS instance: %v\n", err)
-					// mutexWatch.Unlock()
-					continue
+				if !isDeleteInBatch {
+					log.Printf("Pod Event: Deleting VMSS instance with ID: %s\n", instanceID)
+					_, err = vmssClient.Delete(context.TODO(), vmssRG, vmssName, instanceID, nil)
+					if err != nil {
+						log.Printf("Pod Event: Error deleting VMSS instance: %v\n", err)
+						// mutexWatch.Unlock()
+						continue
+					}
 				}
 				mutexWatch.Lock()
 				delete(nodesToDeleteWatch, nodeName) // 删除成功后，从删除队列中移除节点
@@ -186,6 +259,21 @@ func processNodesbyEvents(clientset *kubernetes.Clientset, vmssClient compute.Vi
 			}
 			// }(nodeName)
 		}
+		defer func() {
+			if isDeleteInBatch {
+				if len(vmssInstances) > 0 {
+					log.Printf("Process Pod Event: Batch Deleting VMSS instance with ID: %s\n", vmssInstances)
+					DeleteVMInstances(vmssClient, vmssInstances)
+				}
+			}
+		}()
+
+		defer func() {
+			if r := recover(); r != nil {
+				log.Printf("Process Pod Event: Recovered from panic: %v\n", r)
+			}
+		}()
+
 		// mutexWatch.Unlock()
 	}
 }
@@ -198,16 +286,21 @@ func checkAndDeleteNodes(clientset *kubernetes.Clientset, vmssClient compute.Vir
 	nodepool := os.Getenv("FOCUS_NODEPOOL")
 	namespace := os.Getenv("NAMESPACE")
 	labelSelector := "agentpool=" + nodepool
+	isDeleteInBatchStr := os.Getenv("DELETE_IN_BATCH")
+	isDeleteInBatch, _ := strconv.ParseBool(isDeleteInBatchStr)
+
 	log.Printf("On Scheduled: Starting to collect pods from nodepool %s", labelSelector)
 	// List all nodes
 	nodes, err := clientset.CoreV1().Nodes().List(context.Background(), metav1.ListOptions{
 		LabelSelector: labelSelector,
 	})
+
 	if err != nil {
 		log.Printf("failed to list nodes: %v\n", err)
 		return fmt.Errorf("failed to list nodes: %v", err)
 	}
 	currentCheck := make(map[string]bool)
+	vmssInstances := []string{}
 	for _, node := range nodes.Items {
 		// Check if the node has the specified label
 		// if val, ok := node.Labels[labelKey]; ok && val == labelValue {
@@ -255,15 +348,16 @@ func checkAndDeleteNodes(clientset *kubernetes.Clientset, vmssClient compute.Vir
 				// }
 
 				// Drain the node
-				err = DrainNode(clientset, node.Name)
-				log.Printf("On Scheduled: Drain node: %s Completed \n", node.Name)
+				log.Printf("On Scheduled: Cordon node: %s Completed \n", node.Name)
+				err = CordonNode(clientset, node.Name)
 				if err != nil {
-					log.Printf("failed to drain node %s: %v\n", node.Name, err)
-					return fmt.Errorf("failed to drain node %s: %v", node.Name, err)
+					log.Printf("failed to Cordon node %s: %v\n", node.Name, err)
+					return fmt.Errorf("failed to Cordon node %s: %v", node.Name, err)
 				}
 
 				// Convert AKS node name to VMSS instance ID
 				vmssName, instanceID, err := convertNodeNameToVMSS(node.Name)
+				vmssInstances = append(vmssInstances, instanceID)
 				if err != nil {
 					log.Printf("failed to convert node name to VMSS instance ID: %v\n", err)
 					return fmt.Errorf("failed to convert node name to VMSS instance ID: %v", err)
@@ -275,11 +369,13 @@ func checkAndDeleteNodes(clientset *kubernetes.Clientset, vmssClient compute.Vir
 					return fmt.Errorf("error deleting node %s from Kubernetes: %v", node.Name, err)
 				}
 				// Delete the VMSS instance
-				log.Printf("On Scheduled: Deleting VMSS instance with ID: %s\n", instanceID)
-				_, err = vmssClient.Delete(context.TODO(), vmssRG, vmssName, instanceID, nil)
-				if err != nil {
-					log.Printf("Error deleting VMSS instance: %v\n", err)
-					return fmt.Errorf("error deleting VMSS instance: %v", err)
+				if !isDeleteInBatch {
+					log.Printf("On Scheduled: Deleting VMSS instance with ID: %s\n", instanceID)
+					_, err = vmssClient.Delete(context.TODO(), vmssRG, vmssName, instanceID, nil)
+					if err != nil {
+						log.Printf("Error deleting VMSS instance: %v\n", err)
+						return fmt.Errorf("error deleting VMSS instance: %v", err)
+					}
 				}
 				delete(nodesToCheckCycle, node.Name)
 				// log.Printf("On Scheduled: Deleted VMSS instance %s and node %s\n", instanceID, node.Name)
@@ -300,6 +396,22 @@ func checkAndDeleteNodes(clientset *kubernetes.Clientset, vmssClient compute.Vir
 		}
 
 	}
+	// isDeleteInBatch := os.Getenv("DELETE_IN_BATCH")
+	defer func() {
+		if isDeleteInBatch {
+			if len(vmssInstances) > 0 {
+				log.Printf("On Scheduled: Batch Deleting VMSS instance with ID: %s\n", vmssInstances)
+				DeleteVMInstances(vmssClient, vmssInstances)
+			}
+		}
+	}()
+
+	defer func() {
+		if r := recover(); r != nil {
+			log.Printf("On Scheduled: Recovered from panic: %v\n", r)
+			// 可以在这里执行任何必要的清理操作
+		}
+	}()
 	// nodesToCheckCycle = currentCheck
 	for nodeName := range currentCheck {
 		nodesToCheckCycle[nodeName] = true
@@ -315,6 +427,11 @@ func checkAndDeleteNodes(clientset *kubernetes.Clientset, vmssClient compute.Vir
 	}
 	//print nodesToCheckCycle
 	log.Printf("On Scheduled: nodesToCheckCycle at this round: %v\n", nodesToCheckCycle)
+
+	return nil
+}
+
+func updateFakeNP(aksClient containerservice.AgentPoolsClient, aksRG string, aksClusterName string, fakeNP string) error {
 	is_fakenp_check_flag_str := os.Getenv("FAKENP_FLAG")
 	is_fakenp_check_flag, err := strconv.ParseBool(is_fakenp_check_flag_str)
 	if err != nil {
@@ -346,7 +463,7 @@ func checkAndDeleteNodes(clientset *kubernetes.Clientset, vmssClient compute.Vir
 		if err != nil {
 			log.Fatalf("Failed to update agent pool: %v", err)
 		}
-		log.Printf("Updated agent pool %s with max count %d\n", fakeNP, maxCount)
+		log.Printf("Fake Nodepool: Updated agent pool %s with max count %d\n", fakeNP, maxCount)
 	}
 	return nil
 }
@@ -580,27 +697,37 @@ func isNodeReady(node *corev1.Node) bool {
 	return false
 }
 
-// DrainNode 使用 kubectl 的 drain 包来驱逐节点上的 Pod
-func DrainNode(clientset *kubernetes.Clientset, nodeName string) error {
-	log.Printf("On Scheduled: Draining node: %s\n", nodeName)
-
-	node, err := clientset.CoreV1().Nodes().Get(context.Background(), nodeName, metav1.GetOptions{})
+// CordonNode 使用 kubectl 的 drain 包来驱逐节点上的 Pod
+func CordonNode(clientset *kubernetes.Clientset, nodeName string) error {
+	// log.Printf("On Scheduled: Draining node: %s\n", nodeName)
+	is_cordonFlagStr := os.Getenv("CORDON_FLAG")
+	is_cordonFlag, err := strconv.ParseBool(is_cordonFlagStr)
 	if err != nil {
-		return err
+		log.Printf("Error converting CORDON_FLAG to bool: %s\n", err)
 	}
+	if is_cordonFlag {
+		node, err := clientset.CoreV1().Nodes().Get(context.Background(), nodeName, metav1.GetOptions{})
+		if err != nil {
+			return err
+		}
 
-	drainer := &drain.Helper{
-		Client:              clientset,
-		Force:               true,
-		IgnoreAllDaemonSets: true,
-		DeleteEmptyDirData:  true,
-		GracePeriodSeconds:  -1,
-		Timeout:             30 * time.Second,
-		Out:                 os.Stdout,
-		ErrOut:              os.Stderr,
+		drainer := &drain.Helper{
+			Client:              clientset,
+			Ctx:                 context.Background(),
+			Force:               true,
+			IgnoreAllDaemonSets: true,
+			DeleteEmptyDirData:  true,
+			GracePeriodSeconds:  -1,
+			Timeout:             30 * time.Second,
+			Out:                 os.Stdout,
+			ErrOut:              os.Stderr,
+		}
+
+		// return drain.RunNodeDrain(drainer, node.Name)
+		return drain.RunCordonOrUncordon(drainer, node, true)
+	} else {
+		return nil
 	}
-
-	return drain.RunNodeDrain(drainer, node.Name)
 }
 
 // AddTaintToNode 使用 client-go 库给节点添加 taint
